refactor-compact-tiles * Use same BikeTile on the compact dashboard as on the chrono dashboard. * Move ChronoBikeTile and ChronoBikeTileRow to components. Rename them to just BikeTile and BikeTileRow
