+++ src/components/BikeTile/BikeTile.module.scss
@@ -1,3 +1,3 @@
-.chronoBikeTile {
+.BikeTile {
     height: 100%;
 }
