+++ src/dashboards/Compact/CompactDashboard.tsx
@@ -21,9 +21,9 @@ import { isMobileWeb } from '../../utils/utils'
 import { WeatherTile } from '../../components/WeatherTile/WeatherTile'
 import { QRTile } from '../../components/QRTile/QRTile'
 import { ImageTile } from '../../components/ImageTile/ImageTile'
+import { BikeTile } from '../../components/BikeTile/BikeTile'
 import { FormFactor } from '../../../graphql-generated/mobility-v2'
 import { DepartureTile } from './DepartureTile/DepartureTile'
-import { BikeTile } from './BikeTile/BikeTile'
 import { MapTile } from './MapTile/MapTile'
 import './CompactDashboard.scss'
